+++ app/controller/settings.js
@@ -1,0 +1,1 @@
+"use strict";
